Remove `if` for index in step 2 answer
Because in Python x[n:m] will return _only_ the results from x that are
actually there, not error if m >= len(x), so the check on the index that
was there before was not needed.

I.e., as the (2000-element list) is scanned, the final lines are effectively:

A: x[1996:1999]
B: x[1997:2000]
C: x[1998:2001]
D: x[1999:2002]

The latter three (lines B-D) return fewer values as indexe value 2000+ is not
present, so the indexing/slicing returns fewer results.

Because we're looking for _increases_, this means that we can leave the
index `if` out, as we _know_ that B:D will be smaller as they contain a
strict subset of the values in the previous record (e.g. B is a subset
of A, C is a subet of B etc.)